01_stock_pipeline/requirements.txt
aioice==0.10.2
aiortc==1.14.0
alembic==1.17.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.12.1
appnope @ file:///home/conda/feedstock_root/build_artifacts/appnope_1733332318622/work
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
attrs==25.4.0
av==16.1.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==5.5.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorlog==6.10.1
comm @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_comm_1753453984/work
contourpy==1.3.3
cryptography==46.0.4
curl_cffi==0.13.0
cycler==0.12.1
debugpy @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_6a37he2v_t/croot/debugpy_1736267437603/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
distro==1.9.0
dnspython==2.8.0
et_xmlfile==2.0.0
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1746947292760/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1756729339227/work
fonttools==4.59.2
frozendict==2.4.7
gitdb==4.0.12
GitPython==3.1.46
google-crc32c==1.8.0
graphviz==0.21
gTTS==2.5.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ifaddr==0.2.0
ImageIO==2.37.2
imbalanced-learn==0.14.0
imblearn==0.0
importlib_metadata==7.2.1
ipykernel @ file:///Users/runner/miniforge3/conda-bld/ipykernel_1754352890318/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_ipython_1756474504/work
ipython_pygments_lexers @ file:///home/conda/feedstock_root/build_artifacts/ipython_pygments_lexers_1737123620466/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.6
jiter==0.12.0
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1756321760188/work
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1748333051527/work
kiwisolver==1.4.9
lazy_loader==0.4
lightgbm==4.6.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.6
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mdurl==0.1.2
multitasking==0.0.12
narwhals==2.10.2
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.6
numpy==1.26.4
openai==2.16.0
opencv-python==4.12.0.88
openpyxl==3.1.5
optuna==4.5.0
packaging==23.2
pandas==2.3.2
parso @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_parso_1755974222/work
patsy==1.0.2
peewee==3.19.0
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1733301927746/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==10.4.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1756227402/work
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1756321756983/work
protobuf==4.25.8
psutil @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_10oa1k8l11/croot/psutil_1736367646006/work
psycopg2-binary==2.9.11
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1733302279685/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=92c32ff62b5fd8cf325bec5ab90d7be3d2a8ca8c8a3813ff487a8d2002630d1f
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pyarrow==23.0.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pyee==13.0.0
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
pylibsrtp==1.0.0
pyOpenSSL==25.3.0
pyparsing==3.2.3
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
python-dotenv==1.2.1
pytz==2025.2
PyWavelets @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_72w3dbiyom/croot/pywavelets_1756715672224/work
PyYAML==6.0.3
pyzmq @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_95lsut8ymz/croot/pyzmq_1734709560733/work
referencing==0.37.0
requests==2.32.5
rich==13.9.4
rpds-py==0.28.0
schedule==1.2.2
scikit-image==0.25.2
scikit-learn @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_aew5bgsgui/croot/scikit-learn_1753427393651/work
scipy==1.16.1
seaborn==0.13.2
setuptools==78.1.1
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
sk==0.0.1
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8.3
SQLAlchemy==2.0.44
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
statsmodels==0.14.6
streamlit==1.40.0
streamlit-webrtc==0.64.5
tenacity==8.5.0
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1741878222898/work
tifffile==2025.10.16
toml==0.10.2
tornado @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_63s3osc0_1/croot/tornado_1748956943583/work
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
typing-inspection==0.4.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.6.3
validators==0.35.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
webrtcvad==2.0.10
websockets==16.0
wheel==0.45.1
xgboost @ file:///Users/runner/miniforge3/conda-bld/xgboost-split_1755048557197/work/python-package
yfinance==1.1.0
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1749421620841/work
